requirements.txt
# API Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==23.0.0
python-multipart==0.0.9

# LangChain
langchain==0.3.1
langchain-openai==0.2.1
langchain-pinecone==0.2.0
langchain-community==0.3.1
langchain-core==0.3.6
langchain-text-splitters==0.3.0

# OpenAI
openai==1.51.0

# Pinecone
pinecone-client==5.0.1

# Redis
redis[asyncio]==5.1.0

# PDF parsing
pymupdf==1.24.10

# Config & validation
pydantic==2.9.2
pydantic-settings==2.5.2

# Logging
loguru==0.7.2

# Retry logic
#tenacity==9.0.0
tenacity==8.3.0
# Numpy
numpy==1.26.4

# Rate limiting
slowapi==0.1.9

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
